Rename quadratic dir to factor
