+++ packages/solver/src/solve/solve.ts
@@ -1,5 +1,4 @@
 import * as Semantic from '@math-blocks/semantic';
-import * as Testing from '@math-blocks/testing';
 
 import { divBothSides } from './transforms/div-both-sides';
 import { mulBothSides } from './transforms/mul-both-sides';
@@ -82,8 +81,6 @@ export function solve(
   }
 
   const after = current;
-
-  Testing.print(after); // ?
 
   if (Semantic.util.isNumericRelation(after)) {
     const [left, right] = after.args;
